fix: use getBoundingClientRect() when receiving touch events
This commit fixes an issue for the touch start and touch end.

The X and Y wasn't properly calculated when the canvas was
offset on the page, i.e. not align to the top left.

To fix the issue use canvas.getBoundingClientRect() instead of
canvas.offsetLeft/canvas.offsetTop so that we take the viewport into
consideration.

Refs:
- https://developer.mozilla.org/en-US/docs/Web/API/Element/getBoundingClientRect